+++ packages/vue-client/src/requests.ts
@@ -11,9 +11,9 @@ async function requestImpl(method: string, path: string, json?: object) {
   const headers = new Headers();
   headers.append("Origin", SERVER_ORIGIN); // TODO: Is this necessary?
   headers.append("Content-Type", "application/json");
-  const csfr_token = useStore().csrf_token;
-  if (csfr_token) {
-    headers.append("CSRF-Token", csfr_token);
+  const csrf_token = useStore().csrf_token;
+  if (csrf_token) {
+    headers.append("CSRF-Token", csrf_token);
   }
   const response = await fetch(SERVER_PATH_PREFIX + path, {
     method,
